refactor: extract LanguageParser trait — adapter per language (#189)
Replace 14 free-function pairs (parse_<lang>_imports + resolve_<lang>_import)
with a LanguageParser trait and one adapter struct per language under
src/scanner/parsers/. The registry() function maps file extensions to adapters,
so scan_project() no longer needs a match arm per language.

Adding a 15th language now requires only: (a) one new adapter file, (b) one line
in parsers/mod.rs::registry(). All 195 existing tests pass; clippy clean.
